Version 1.8.6
## Version 1.8.6 - October 17, 2016

- **Improved:** Updated GleamTech.Core.

- **Improved:** Image cache will now group cached files for the same
inputs in subfolders. Also the subfolder name
will contain the original file name so it will be easier to locate the
corresponding cached files for an input.
The old cached files will be arranged into the new structure
automatically one by one when they are accessed so
no action is required for the developer.