Version 1.9.4
## Version 1.9.4 - February 13, 2017

- **Fixed:** Setting LicenseKey via code, did not remove trial
restrictions if application was not restarted,
e.g. if the code is changed in .aspx or or .asax markup this would not
cause the whole application recompile & restart.
Setting LicenseKey via appSettings in web.config worked because this
already forces application restart.

- **Fixed:** Setting ImageUltimateWebConfiguration.CacheMaxDays to a
value 0 or less will not cause error during
caching a file. Current cache subfolder will be excluded from cache trim
operation
so during caching a file, related files for cache key will not be
deleted mistakenly.